implement VirtualFolderDbId discoverChildren added MusicBrainzAlbumFolder
