+++ src/main/java/net/pms/network/mediaserver/handlers/SearchRequestHandler.java
@@ -16,7 +16,6 @@
  */
 package net.pms.network.mediaserver.handlers;
 
-import java.io.File;
 import java.sql.Connection;
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -25,22 +24,21 @@ import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
-import java.util.concurrent.atomic.AtomicInteger;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import net.pms.database.MediaDatabase;
 import net.pms.database.MediaTableAudioMetadata;
 import net.pms.dlna.DidlHelper;
 import net.pms.formats.Format;
-import net.pms.media.audio.metadata.MusicBrainzAlbum;
 import net.pms.network.mediaserver.HTTPXMLHelper;
 import net.pms.network.mediaserver.handlers.message.SearchRequest;
 import net.pms.renderers.Renderer;
 import net.pms.store.DbIdMediaType;
 import net.pms.store.DbIdResourceLocator;
 import net.pms.store.DbIdTypeAndIdent;
+import net.pms.store.MediaStoreIds;
 import net.pms.store.StoreResource;
-import net.pms.store.container.VirtualFolderDbId;
+import net.pms.store.container.MusicBrainzAlbumFolder;
 import net.pms.store.container.VirtualFolderDbIdNamed;
 import org.apache.commons.io.FilenameUtils;
 import org.apache.commons.lang3.StringUtils;
@@ -69,8 +67,6 @@ public class SearchRequestHandler {
 
 	private static final Pattern TOKENIZER_PATTERN = Pattern.compile(
 		"(?<property>((\\bdc\\b)|(\\bupnp\\b)):[A-Za-z@\\[\\]\"=]+)\\s+(?<op>[A-Za-z=!<>]+)\\s+\"(?<val>.*?)\"", Pattern.CASE_INSENSITIVE);
-
-	private final AtomicInteger updateID = new AtomicInteger(1);
 
 	public static DbIdMediaType getRequestType(String searchCriteria) {
 		LOGGER.debug("search criteria : {}", searchCriteria);
@@ -127,22 +123,14 @@ public class SearchRequestHandler {
 
 		int totalMatches = getLibraryResourceCountFromSQL(convertToCountSql(requestMessage.getSearchCriteria(), requestType));
 
-		VirtualFolderDbId folder = new VirtualFolderDbId(renderer, "SearchResult", new DbIdTypeAndIdent(requestType, ""));
 		String sqlFiles = convertToFilesSql(requestMessage, requestType);
 		for (StoreResource resource : getLibraryResourceFromSQL(renderer, sqlFiles, requestType)) {
-			folder.addChild(resource);
-		}
-
-		folder.discoverChildren();
-		for (StoreResource uf : folder.getChildren()) {
 			numberReturned++;
-			uf.resolve();
-			uf.setFakeParentId("0");
-			dlnaItems.append(DidlHelper.getDidlString(uf));
+			dlnaItems.append(DidlHelper.getDidlString(resource));
 		}
 
 		// Build response message
-		StringBuilder response = buildEnvelope(numberReturned, totalMatches, updateID.getAndIncrement(), dlnaItems);
+		StringBuilder response = buildEnvelope(numberReturned, totalMatches, MediaStoreIds.getSystemUpdateId().getValue(), dlnaItems);
 		return createResponse(response.toString());
 	}
 
@@ -489,7 +477,7 @@ public class SearchRequestHandler {
 	 * @return
 	 */
 	public static List<StoreResource> getLibraryResourceFromSQL(Renderer renderer, String query, DbIdMediaType type) {
-		ArrayList<StoreResource> filesList = new ArrayList<>();
+		ArrayList<StoreResource> result = new ArrayList<>();
 
 		if (LOGGER.isTraceEnabled()) {
 			LOGGER.trace(String.format("SQL %s : %s", type.dbidPrefix, query));
@@ -507,38 +495,40 @@ public class SearchRequestHandler {
 								case TYPE_ALBUM -> {
 									String mbid = resultSet.getString("MBID_RECORD");
 									if (StringUtils.isAllBlank(mbid)) {
-										filesList.add(new VirtualFolderDbIdNamed(renderer, filenameField, new DbIdTypeAndIdent(type, filenameField)));
+										result.add(new VirtualFolderDbIdNamed(renderer, filenameField, new DbIdTypeAndIdent(type, filenameField)));
 									} else {
 										if (!foundMbidAlbums.contains(mbid)) {
-											VirtualFolderDbIdNamed albumFolder = new VirtualFolderDbIdNamed(renderer, filenameField,
-												new DbIdTypeAndIdent(DbIdMediaType.TYPE_MUSICBRAINZ_RECORDID, mbid));
-											MusicBrainzAlbum album = new MusicBrainzAlbum(resultSet.getString("MBID_RECORD"),
+											MusicBrainzAlbumFolder albumFolder = new MusicBrainzAlbumFolder(renderer, filenameField, mbid,
 												resultSet.getString("album"), resultSet.getString("artist"), resultSet.getInt("media_year"),
 												resultSet.getString("genre"));
-											DbIdResourceLocator.appendAlbumInformation(album, albumFolder);
-											filesList.add(albumFolder);
+											result.add(albumFolder);
 											foundMbidAlbums.add(mbid);
 										}
 									}
 								}
-								case TYPE_PERSON -> filesList
-									.add(new VirtualFolderDbIdNamed(renderer, filenameField, new DbIdTypeAndIdent(type, filenameField)));
-								case TYPE_PERSON_COMPOSER -> filesList.add(new VirtualFolderDbIdNamed(renderer, filenameField,
-									new DbIdTypeAndIdent(type, DbIdMediaType.PERSON_COMPOSER_PREFIX + filenameField)));
-								case TYPE_PERSON_CONDUCTOR -> filesList.add(new VirtualFolderDbIdNamed(renderer, filenameField,
-									new DbIdTypeAndIdent(type, DbIdMediaType.PERSON_CONDUCTOR_PREFIX + filenameField)));
-								case TYPE_PERSON_ALBUMARTIST -> filesList.add(new VirtualFolderDbIdNamed(renderer, filenameField,
-									new DbIdTypeAndIdent(type, DbIdMediaType.PERSON_ALBUMARTIST_PREFIX + filenameField)));
-								case TYPE_PLAYLIST -> filesList
-									.add(new VirtualFolderDbIdNamed(renderer, filenameField, new DbIdTypeAndIdent(type, resultSet.getString("FID"))));
+								case TYPE_PERSON, TYPE_PERSON_COMPOSER, TYPE_PERSON_CONDUCTOR, TYPE_PERSON_ALBUMARTIST -> {
+									DbIdTypeAndIdent typeIdent = new DbIdTypeAndIdent(type, filenameField);
+									StoreResource resource = DbIdResourceLocator.getLibraryResourceByDbTypeIdent(renderer, typeIdent);
+									if (resource != null) {
+										result.add(resource);
+									}
+								}
+								case TYPE_PLAYLIST -> {
+									String realFileName = resultSet.getString("FILENAME");
+									if (realFileName != null) {
+										StoreResource res = DbIdResourceLocator.getLibraryResourcePlaylist(renderer, realFileName);
+										if (res != null) {
+											result.add(res);
+										}
+									}
+								}
 								default -> {
 									String realFileName = resultSet.getString("FILENAME");
 									if (realFileName != null) {
-										StoreResource item = DbIdResourceLocator.createResourceFromFile(
-											renderer,
-											new File(realFileName),
-											new DbIdTypeAndIdent(type, resultSet.getString("FID")));
-										filesList.add(item);
+										StoreResource res = DbIdResourceLocator.getLibraryResourceRealFile(renderer, realFileName);
+										if (res != null) {
+											result.add(res);
+										}
 									}
 								}
 
@@ -552,7 +542,7 @@ public class SearchRequestHandler {
 		} finally {
 			MediaDatabase.close(connection);
 		}
-		return filesList;
+		return result;
 	}
 
 	private static String extractDisplayName(ResultSet resultSet, DbIdMediaType type) throws SQLException {
@@ -583,7 +573,7 @@ public class SearchRequestHandler {
 		return response;
 	}
 
-	private static StringBuilder buildEnvelope(int foundNumberReturned, int totalMatches, int updateID, StringBuilder dlnaItems) {
+	private static StringBuilder buildEnvelope(int foundNumberReturned, int totalMatches, long updateID, StringBuilder dlnaItems) {
 		StringBuilder response = new StringBuilder();
 		response.append(HTTPXMLHelper.SEARCHRESPONSE_HEADER);
 		response.append(CRLF);
